chore code improvements
