+++ example/lib/main.dart
@@ -5,13 +5,11 @@ import 'dart:io';
 
 import 'package:app_tracking_transparency/app_tracking_transparency.dart';
 import 'package:axeptio_sdk/axeptio_sdk.dart';
-import 'package:axeptio_sdk/events/event_listener.dart';
 import 'package:axeptio_sdk_example/shared_preferences_dialog.dart';
 import 'package:axeptio_sdk_example/tokendialog.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:google_mobile_ads/google_mobile_ads.dart';
-import 'package:axeptio_sdk/model/axeptio_service.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
