+++ lib/src/channel/channel.dart
@@ -1,0 +1,3 @@
+export 'axeptio.dart';
+export 'axeptio_sdk_method_channel.dart';
+export 'axeptio_sdk_platform_interface.dart';
