+++ lib/src/events/event_listener.dart
@@ -1,4 +1,4 @@
-import 'package:axeptio_sdk/model/consents_v2.dart';
+import 'package:axeptio_sdk/src/model/consents_v2.dart';
 
 class AxeptioEventListener {
   dynamic Function() onPopupClosedEvent = () {};
