+++ lib/src/model/model.dart
@@ -1,0 +1,2 @@
+export 'axeptio_service.dart';
+export 'consents_v2.dart';
