+++ test/axeptio_sdk_test.dart
@@ -1,8 +1,4 @@
 import 'package:axeptio_sdk/axeptio_sdk.dart';
-import 'package:axeptio_sdk/axeptio_sdk_method_channel.dart';
-import 'package:axeptio_sdk/axeptio_sdk_platform_interface.dart';
-import 'package:axeptio_sdk/events/event_listener.dart';
-import 'package:axeptio_sdk/model/axeptio_service.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
@@ -73,6 +69,7 @@ class MockAxeptioSdkPlatform
     // TODO: implement removeEventListener
     throw UnimplementedError();
   }
+  
 }
 
 void main() {
